Basic external hashing solution for cogrouped rdd.
We statically bucket into 8192 buckets, which is based on the estimation
that misconfiguration for the number of partitions is not going to be off
by more than a factor of ~10k. (The exact number is configurable.)

This code is 100% test code, not intended for any sort of real production
purposes.